requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==2.2
django-allauth==0.44.0
django-braces==1.14.0
django-import-export==2.5.0
django-model-utils==4.1.1
django-private-chat==0.3.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
isort==5.7.0
itemadapter==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
ndg-httpsclient==0.5.1
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
pandas==1.2.1
parsel==1.6.0
Pillow==8.1.0
psycopg2==2.8.6
pusher==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.1
pylint==2.6.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.2
virtualenv==20.4.0
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.2.0
